+++ com.unity.render-pipelines.high-definition/Editor/Material/UIBlocks/DetailInputsUIBlock.cs
@@ -29,8 +29,7 @@ namespace UnityEditor.Rendering.HighDefinition
 
         internal class Styles
         {
-            public const string header = "Detail Inputs";
-
+            public static GUIContent header { get; } = EditorGUIUtility.TrTextContent("Detail Inputs");
             public static GUIContent UVDetailMappingText = new GUIContent("Detail UV Mapping", "");
             public static GUIContent detailMapNormalText = new GUIContent("Detail Map", "Specifies the Detail Map albedo (R) Normal map y-axis (G) Smoothness (B) Normal map x-axis (A) - Neutral value is (0.5, 0.5, 0.5, 0.5)");
             public static GUIContent detailAlbedoScaleText = new GUIContent("Detail Albedo Scale", "Controls the scale factor for the Detail Map's Albedo.");
@@ -60,7 +59,6 @@ namespace UnityEditor.Rendering.HighDefinition
         MaterialProperty displacementMode = null;
         const string kDisplacementMode = "_DisplacementMode";
 
-        ExpandableBit  m_ExpandableBit;
         Features    m_Features;
         int         m_LayerIndex;
         int         m_LayerCount;
@@ -77,8 +75,8 @@ namespace UnityEditor.Rendering.HighDefinition
         /// <param name="features">Features of the block.</param>
         /// <param name="dotColor">Subheader dot color. See Layered Lit UI subheader for more info.</param>
         public DetailInputsUIBlock(ExpandableBit expandableBit, int layerCount = 1, int layerIndex = 0, Features features = Features.All, Color dotColor = default(Color))
+            : base(expandableBit, Styles.header)
         {
-            m_ExpandableBit = expandableBit;
             m_Features = features;
             m_LayerIndex = layerIndex;
             m_LayerCount = layerCount;
@@ -102,23 +100,19 @@ namespace UnityEditor.Rendering.HighDefinition
         }
 
         /// <summary>
-        /// Renders the properties in the block.
+        /// Property that specifies if the scope is a subheader
         /// </summary>
-        public override void OnGUI()
-        {
-            bool subHeader = (m_Features & Features.SubHeader) != 0;
-
-            using (var header = new MaterialHeaderScope(Styles.header, (uint)m_ExpandableBit, materialEditor, subHeader: subHeader, colorDot: m_DotColor))
-            {
-                if (header.expanded)
-                    DrawDetailsGUI();
-            }
-        }
+        protected override bool isSubHeader => (m_Features & Features.SubHeader) != 0;
 
         /// <summary>
-        /// Draw the Details GUI.
+        /// Dot color for material scope
         /// </summary>
-        protected void DrawDetailsGUI()
+        protected override Color dotColor => m_DotColor;
+
+        /// <summary>
+        /// Renders the properties in the block.
+        /// </summary>
+        protected override void OnGUIInternal()
         {
             UVBaseMapping uvBaseMapping = (UVBaseMapping)UVBase[m_LayerIndex].floatValue;
             float X, Y, Z, W;