+++ com.unity.render-pipelines.high-definition/Editor/Material/UIBlocks/DistortionUIBlock.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.Rendering.HighDefinition;
 
 namespace UnityEditor.Rendering.HighDefinition
 {
@@ -12,6 +9,7 @@ namespace UnityEditor.Rendering.HighDefinition
     {
         internal static class Styles
         {
+            public static GUIContent header { get; } = EditorGUIUtility.TrTextContent("Distortion");
             public static GUIContent distortionEnableText = new GUIContent("Distortion", "When enabled, HDRP processes distortion for this Material.");
             public static GUIContent distortionOnlyText = new GUIContent("Distortion Only", "When enabled, HDRP only uses this Material to render distortion.");
             public static GUIContent distortionDepthTestText = new GUIContent("Distortion Depth Test", "When enabled, HDRP calculates a depth test for distortion.");