+++ com.unity.render-pipelines.high-definition/Editor/Material/UIBlocks/ShaderGraphUIBlock.cs
@@ -1,9 +1,7 @@
 using System;
-using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 using UnityEngine.Rendering.HighDefinition;
-using System.Linq;
-
 // Include material common properties names
 using static UnityEngine.Rendering.HighDefinition.HDMaterialProperties;
 
@@ -35,10 +33,9 @@ namespace UnityEditor.Rendering.HighDefinition
 
         internal static class Styles
         {
-            public const string header = "Exposed Properties";
+            public static GUIContent header { get; } = EditorGUIUtility.TrTextContent("Exposed Properties");
         }
 
-        ExpandableBit  m_ExpandableBit;
         Features    m_Features;
 
         /// <summary>
@@ -47,8 +44,8 @@ namespace UnityEditor.Rendering.HighDefinition
         /// <param name="expandableBit">Bit index used to store the foldout state.</param>
         /// <param name="features">Features enabled in the block.</param>
         public ShaderGraphUIBlock(ExpandableBit expandableBit = ExpandableBit.ShaderGraph, Features features = Features.All)
+            : base(expandableBit, Styles.header)
         {
-            m_ExpandableBit = expandableBit;
             m_Features = features;
         }
 
@@ -60,13 +57,24 @@ namespace UnityEditor.Rendering.HighDefinition
         /// <summary>
         /// Renders the properties in the block.
         /// </summary>
-        public override void OnGUI()
+        protected override void OnGUIInternal()
         {
-            using (var header = new MaterialHeaderScope(Styles.header, (uint)m_ExpandableBit, materialEditor))
+            // Filter out properties we don't want to draw:
+            if ((m_Features & Features.ExposedProperties) != 0)
+                PropertiesDefaultGUI(properties);
+
+            // If we change a property in a shadergraph, we trigger a material keyword reset
+            if (CheckPropertyChanged(properties))
             {
-                if (header.expanded)
-                    DrawShaderGraphGUI();
+                foreach (var material in materials)
+                    HDShaderUtils.ResetMaterialKeywords(material);
             }
+
+            if ((m_Features & Features.DiffusionProfileAsset) != 0)
+                DrawDiffusionProfileUI();
+
+            if ((m_Features & Features.ShadowMatte) != 0 && materials.All(m => m.HasProperty(kShadowMatteFilter)))
+                DrawShadowMatteToggle();
         }
 
         MaterialProperty[] oldProperties;
@@ -109,26 +117,6 @@ namespace UnityEditor.Rendering.HighDefinition
             oldProperties = properties;
 
             return propertyChanged;
-        }
-
-        void DrawShaderGraphGUI()
-        {
-            // Filter out properties we don't want to draw:
-            if ((m_Features & Features.ExposedProperties) != 0)
-                PropertiesDefaultGUI(properties);
-
-            // If we change a property in a shadergraph, we trigger a material keyword reset
-            if (CheckPropertyChanged(properties))
-            {
-                foreach (var material in materials)
-                    HDShaderUtils.ResetMaterialKeywords(material);
-            }
-
-            if ((m_Features & Features.DiffusionProfileAsset) != 0)
-                DrawDiffusionProfileUI();
-
-            if ((m_Features & Features.ShadowMatte) != 0 && materials.All(m => m.HasProperty(kShadowMatteFilter)))
-                DrawShadowMatteToggle();
         }
 
         /// <summary>